+++ EnemyMovement.h
@@ -9,7 +9,7 @@ struct EnemyMovement {
     bool movingRight = true;
 
     static void InitEnemyMovement(EnemyMovement& move);
-    static void UpdateEnemyPatrol(EnemyGameObject* enemy, f32 dt);
+    static void UpdateEnemyPatrol(EnemyGameObject* enemy, f64 dt);
 
     // A* Pathfinding
     static std::vector<AEVec2> FindPath(AEVec2 start, AEVec2 target);
